Crate mas_tower

Structs

DurationRecorderFuture

The future returned by the DurationRecorderService.

DurationRecorderLayer

A Layer that records the duration of requests in milliseconds.

DurationRecorderService

A middleware that records the duration of requests in milliseconds.

FnWrapper

A wrapper around a function that can be used to generate a key-value pair, make or enrich spans.

IdentityLayer

A no-op layer that has the request type bound.

IdentityService

A no-op service that has the request type bound.

InFlightCounterLayer

A Layer that records the number of in-flight requests.

InFlightCounterService

A middleware that records the number of in-flight requests.

InFlightFuture

The future returned by InFlightCounterService

KV

A simple static key-value pair.

TraceContextLayer

A Layer that adds a trace context to the request.

TraceContextService

A Service that adds a trace context to the request.

TraceFuture

TraceLayer

TraceService

Traits

EnrichSpan

A trait for enriching a span with information a structure.

MakeSpan

A trait for creating a span for a request.

MetricsAttributes

Make metrics attributes from a type.

Functions

enrich_span_fn

Enrich span from a function.

make_span_fn

Make span from a function.

metrics_attributes_fn